Move update modules to a separate directory
There are enough update files that they should really just be in their own directory.
