+++ app/update/result_test.py
@@ -2,14 +2,14 @@ import datetime
 import io
 import unittest
 
-import update_result
+import update.result
 
 
 class UpdateResultTest(unittest.TestCase):
 
     def test_reads_correct_values_for_successful_result(self):
         self.assertEqual(
-            update_result.Result(
+            update.result.Result(
                 error='',
                 timestamp=datetime.datetime(2021,
                                             2,
@@ -19,7 +19,7 @@ class UpdateResultTest(unittest.TestCase):
                                             35,
                                             tzinfo=datetime.timezone.utc),
             ),
-            update_result.read(
+            update.result.read(
                 io.StringIO("""
 {
   "error": "",
@@ -29,7 +29,7 @@ class UpdateResultTest(unittest.TestCase):
 
     def test_reads_correct_values_for_failed_result(self):
         self.assertEqual(
-            update_result.Result(
+            update.result.Result(
                 error='dummy update error',
                 timestamp=datetime.datetime(2021,
                                             2,
@@ -39,7 +39,7 @@ class UpdateResultTest(unittest.TestCase):
                                             35,
                                             tzinfo=datetime.timezone.utc),
             ),
-            update_result.read(
+            update.result.read(
                 io.StringIO("""
 {
   "error": "dummy update error",
@@ -49,15 +49,15 @@ class UpdateResultTest(unittest.TestCase):
 
     def test_reads_default_values_for_empty_dict(self):
         self.assertEqual(
-            update_result.Result(
+            update.result.Result(
                 error='',
                 timestamp=datetime.datetime.utcfromtimestamp(0),
-            ), update_result.read(io.StringIO('{}')))
+            ), update.result.read(io.StringIO('{}')))
 
     def test_writes_success_result_accurately(self):
         mock_file = io.StringIO()
-        update_result.write(
-            update_result.Result(
+        update.result.write(
+            update.result.Result(
                 error='',
                 timestamp=datetime.datetime(2021,
                                             2,
@@ -72,8 +72,8 @@ class UpdateResultTest(unittest.TestCase):
 
     def test_writes_error_result_accurately(self):
         mock_file = io.StringIO()
-        update_result.write(
-            update_result.Result(
+        update.result.write(
+            update.result.Result(
                 error='dummy update error',
                 timestamp=datetime.datetime(2021,
                                             2,
