+++ scripts/update-service
@@ -10,9 +10,9 @@ import logging
 import os
 import subprocess
 
-import update
-import update_result
-import update_result_reader
+import update.result
+import update.result_reader
+import update.update
 import utc
 
 logger = logging.getLogger(__name__)
@@ -37,7 +37,7 @@ def perform_update():
 
 
 def run_update_script():
-    result = update_result.Result(
+    result = update.result.Result(
         error='The update did not complete',
         timestamp=datetime.datetime.utcfromtimestamp(0))
     try:
@@ -59,11 +59,11 @@ def run_update_script():
 
 
 def write_result(result):
-    result_path = update_result_reader.get_result_path(result.timestamp)
+    result_path = update.result_reader.get_result_path(result.timestamp)
     os.makedirs(os.path.dirname(result_path), exist_ok=True)
     with open(result_path, 'w') as result_file:
         logger.info('Writing result file to %s', result_path)
-        update_result.write(result, result_file)
+        update.result.write(result, result_file)
 
 
 def main(_):
